login working as intended
Learned an odd lesson today. Login logic was working as intended however the data wasn't formatted properly. I finally realized after debugging that i manually entered the data i was testing and because of the way i did that the password wasn't hashed causing the code to throw errors. Code was working as intended just took me some time to figure out the data was the issue.